msm: gsi: change some types in gsi_chan_props
Change the ring_len field to have type u32, and the ring_base_addr
to have type dma_addr_t in the gsi_chan_props structure.  This
is in preparation for the next patch.

+++ drivers/platform/msm/ipa_lite/gsi_dbg.c
@@ -134,7 +134,7 @@ static ssize_t gsi_dump_ch(struct file *file,
 	char *sptr, *token;
 	uint32_t val;
 	struct gsi_chan_ctx *ctx;
-	uint16_t i;
+	u32 i;
 
 	if (sizeof(dbg_buff) < count + 1)
 		return -EFAULT;
